requirements.txt
fastapi
uvicorn
python-dotenv
requests
streamlit
langchain
langchain-openai
langsmith
openai
pydantic